+++ src/middlewares/logger.middleware.js
@@ -1,4 +1,5 @@
 const logRequests = async (req, res, next) => {
+  // TODO: Implement Logger service
   console.log(`Request IP: ${req.ip} to ${req.path}`);
   next();
 };