+++ src/middlewares/rate-limiters.middleware.js
@@ -1,45 +1,61 @@
 const rateLimit = require('express-rate-limit');
 const RedisStore = require('rate-limit-redis');
 const { getClient } = require('../db/redis/init-redis');
+const { insertRequest } = require('../db/cassandra/statistics.repository');
 
 const DEFAULT_WINDOW_TIME = 60 * 60 * 1000; // 1 hour window default
 const DEFAULT_MAX_REQUESTS = 1000; // 1000 requests by window default
 
 const redisClient = getClient();
 
+const createRateLimiter = (
+  keyGenerator,
+  windowMs = DEFAULT_WINDOW_TIME,
+  max = DEFAULT_MAX_REQUESTS,
+  standardHeaders = true,
+  legacyHeaders = false
+) => {
+  return rateLimit({
+    windowMs,
+    max, // Number of requests by window
+    standardHeaders, // Return rate limit info in the `RateLimit-*` headers
+    legacyHeaders, // Disable the `X-RateLimit-*` headers
+    keyGenerator, // Key to identify limiter and count limits
+    store: new RedisStore({
+      sendCommand: (...args) => redisClient.call(...args),
+    }),
+    handler: (req, res, next, options) => {
+      // Handle response when client reached limit
+      const { path, ip, method } = req;
+      insertRequest(path, ip, method, options.statusCode);
+      return res.status(options.statusCode).send(options.message);
+    },
+    message: {
+      error: 'Too many requests, please try again later.',
+      statusCode: 429,
+    },
+  });
+};
+
 // Rate limits by IP Adress
-const ipRateLimiter = rateLimit({
-  windowMs: process.env.IP_RATE_LIMIT_WINDOW_MINUTES * 60 * 1000 || DEFAULT_WINDOW_TIME,
-  max: process.env.IP_RATE_LIMIT_MAX_REQUESTS || DEFAULT_MAX_REQUESTS, // Number of requests by window
-  standardHeaders: true, // Return rate limit info in the `RateLimit-*` headers
-  legacyHeaders: false, // Disable the `X-RateLimit-*` headers
-  store: new RedisStore({
-    sendCommand: (...args) => redisClient.call(...args),
-  }),
-});
+const ipRateLimiter = createRateLimiter(
+  (req) => req.ip,
+  +process.env.IP_RATE_LIMIT_WINDOW_MINUTES * 60 * 1000,
+  +process.env.IP_RATE_LIMIT_MAX_REQUESTS
+);
 
 // Rate limits by path (include path params, not includes query params).
-const pathRateLimiter = rateLimit({
-  windowMs: process.env.PATH_RATE_LIMIT_WINDOW_MINUTES * 60 * 1000 || DEFAULT_WINDOW_TIME,
-  max: process.env.PATH_RATE_LIMIT_MAX_REQUESTS || DEFAULT_MAX_REQUESTS,
-  standardHeaders: true,
-  legacyHeaders: false,
-  keyGenerator: (req) => req.path,
-  store: new RedisStore({
-    sendCommand: (...args) => redisClient.call(...args),
-  }),
-});
+const pathRateLimiter = createRateLimiter(
+  (req) => req.path,
+  +process.env.PATH_RATE_LIMIT_WINDOW_MINUTES * 60 * 1000,
+  +process.env.PATH_RATE_LIMIT_MAX_REQUESTS
+);
 
 // Rate limits by IP Address and path
-const ipAndPathRateLimiter = rateLimit({
-  windowMs: process.env.IP_PATH_RATE_LIMIT_WINDOW_MINUTES * 60 * 1000 || DEFAULT_WINDOW_TIME,
-  max: process.env.IP_PATH_RATE_LIMIT_MAX_REQUESTS || DEFAULT_MAX_REQUESTS,
-  standardHeaders: true,
-  legacyHeaders: false,
-  keyGenerator: (req) => `${req.ip}:${req.path}`,
-  store: new RedisStore({
-    sendCommand: (...args) => redisClient.call(...args),
-  }),
-});
+const ipAndPathRateLimiter = createRateLimiter(
+  (req) => `${req.ip}:${req.path}`,
+  +process.env.IP_PATH_RATE_LIMIT_WINDOW_MINUTES * 60 * 1000,
+  +process.env.IP_PATH_RATE_LIMIT_MAX_REQUESTS
+);
 
 module.exports = { ipRateLimiter, pathRateLimiter, ipAndPathRateLimiter };